server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
browser-use==0.1.40
jsonpath-ng==1.5.0
browserforge==1.2.3
camoufox==0.4.11
beautifulsoup4==4.13.3
jsonschema==4.21.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
cryptography==44.0.2
deprecation==2.1.0
fastapi==0.115.12
frozenlist==1.5.0
geoip2==5.0.1
google-generativeai==0.3.2
gotrue==2.12.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
hypothesis==6.130.7
idna==3.10
iniconfig==2.1.0
language-tags==1.2.0
lxml==5.3.1
maxminddb==2.6.3
multidict==6.3.0
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyee==12.1.1
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
requests==2.32.3
scipy==1.15.2
screeninfo==0.8.1
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3
Appium-Python-Client==3.2.1
selenium==4.18.1
python-multipart==0.0.9
